requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-cors-headers==4.1.0
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.4
packaging==23.1
Pillow==9.5.0
protobuf==3.20.3
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.3
